Added routing and DB query for userTopSongs (#8)
Added the User top songs table, and created DB queries, to return the
user's top songs, and also update the {1,2,3,4}^th position of a users
favorite song with a new song or create it, if something does not exist.
Added a route.ts in the API/userTopSongs folder to provide API access to
these DB queries.

- New feature (non-breaking change which adds functionality)

Local testing, ensuring that DB is updated and produced results match DB
history

-  ✅  My code follows the style guidelines of this project
-  ✅  I have performed a self-review of my code
-  ✅  I have commented my code, particularly in hard-to-understand areas
-  ✅  I have made corresponding changes to the documentation
-  ✅ My changes generate no new warnings
- ✅ I have added tests that prove my fix is effective or that my feature
works
-  ✅  New and existing unit tests pass locally with my changes
- ✅ Any dependent changes have been merged and published in downstream
modules