Ticketmaster Integration (#12)
# Description

Added Ticketmaster API function, that searches for events based on an
artist name, will return at most 10 upcoming events, may return fewer or
none if nothing is found.

## Type of change
- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- This change requires a documentation update

# How Has This Been Tested?
Query of the ticketmaster API works locally, will need to provide the
developer key in the .env file and assign to the variable
TICKET_MASTER_API_KEY to be able to test locally.

# Checklist:

-  ✅ My code follows the style guidelines of this project
-  ✅ I have performed a self-review of my code
-  ✅ I have commented my code, particularly in hard-to-understand areas
-  ✅ I have made corresponding changes to the documentation
-  ✅ My changes generate no new warnings
- ✅ I have added tests that prove my fix is effective or that my feature
works
-  ✅ New and existing unit tests pass locally with my changes
- ✅ Any dependent changes have been merged and published in downstream
modules