Create public user profile page (#20)
# Description

<!-- Please include a summary of the changes and the related issue.
Please also include relevant motivation and context. List any
dependencies that are required for this change. -->

Added a profile page that is accessible under /users/[user_id], where
user_id is the unique user id of each user. We use the user_id in the
path, as the user_id is unique to each user, whereas the username is not
unique. Has tabs for a overview page, reviews page, and concerts page
(stretch feature). Includes some additional changes to the
getUserSongReviews() function to be compatible with the new Review
components and the ReviewProps.

<img width="975" alt="image"
src="https://github.com/user-attachments/assets/0407448c-c559-477d-8cc2-21bd79393eb3"
/>

Not included in this PR: update profile button (if this is your
profile), display of top albums, UI to update top albums, also
implementing liking reviews on the frontend.

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

Tested locally on myself and some other users in the users database. 

# Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules

---------

Co-authored-by: Amanda Wang <52474776+theamandawang@users.noreply.github.com>