Bug Fix for User Auth errors with token expiry (#28)
# Description

This modifies all the checks in session to check for
`session.user.expires_at` which is the expected expiry we get from
Spotify, not the expected expiry of the token itself in the browser,
which is what our code previously did.


Fixes #21


## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration

Verified locally, as well as updated the `Home.test.tsx` and
`Search.test.tsx` files.

# Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my code
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules